refactor(proposal): Refactor of route middlewares and proposal routes.
