+++ app/Http/Controllers/Quotes/Proposal/ReviewQuoteProposalController.php
@@ -3,15 +3,20 @@
 namespace App\Http\Controllers\Quotes\Proposal;
 
 use App\Constants\QuoteProposalTypeConstant;
+use App\Constants\UserTypesAbilitiesConstant;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Quotes\Proposal\CreateQuoteProposalRequest;
 use App\Models\QuoteProposal;
 use Illuminate\Http\JsonResponse;
 
-class ReviewQuoteProposalcontroller extends Controller
+class ReviewQuoteProposalController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware(['auth:sanctum', 'abilities:' . UserTypesAbilitiesConstant::ADMIN]);
+    }
 
-    public function __invoke(CreateQuoteProposalRequest $request, QuoteProposal $quote_proposal) : JsonResponse
+    public function __invoke(CreateQuoteProposalRequest $request, QuoteProposal $quote_proposal): JsonResponse
     {
         $data = $request->validated();
